pyproject.toml
[project]
name = "cancer-cell-portal-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.16",
    "anndata>=0.11.4",
    "annsql>=0.9.2",
    "cellxgene-census>=1.17.0",
    "cellxgene-schema>=5.3.2",
    "jupyter>=1.1.1",
    "jupyterlab-vim>=4.1.4",
    "numcodecs<0.16.0",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.40",
    "supabase>=2.15.0",
    "zarr[remote]<3",
]

[dependency-groups]
dev = [
    "pyfakefs>=5.8.0",
    "pytest>=8.3.5",
    "pytest-xdist>=3.6.1",
]
